Final Cleanup!
Everything is working (I think... should have really wrote some
tests for this thing) and I've cleaned up the README.md as well as
gem versions (let's hope that didn't break anything) and yeah...

Other than some general testing and fixing, eveything is looking
pretty good. Like I mention in the README.md, this isn't production
code. Just code that I put together as an example for a blog post.

So, if you're reading this post, be aware that this is not, I repeat
NOT, production code. Please do not use it in production. Thanks! :-)

+ Updated gem versions
+ Updated README
+ Updated config.ru to just run the app already
+ Updated app.rb to read fences
+ Updated mongo_core to read fences
+ deleted a useless method from Geofence.rb